requirements.txt
python-dotenv
spotipy
openpyxl